requirements.txt
colorama>=0.4.4
pyyaml>=6.0